+++ todo/src/reducers/reducer.js
@@ -1,13 +1,11 @@
-import React, {useState, useReducer} from 'react';
-
-const initialState = {
+export const initialState = [{
     item: 'Learn about reducers',
     completed: false,
     id: 3892987589,
 
-}
+}];
 
-const reducer = {state, action} => {
+const reducer = (state, action) => {
     switch (action.type) {
         case 'ADD_TODO':
             return {
@@ -17,48 +15,18 @@ const reducer = {state, action} => {
             case 'MARK_TODO':
                 return {
                     ...state,
-                    editing: true
+                    editing: false,
+                    title: action.payload
+                }
+                case'CLEAR_TODO':
+                return {
+
+                ...state,
+                editing: false
                 }
     }
 }
 
-const addTodo = () => {
-    const [state, dispatch] = useReducer(reducer, initialState)
-    const [newTodo, setNewTodo] = useState('');
-    const handleChanges = e => {
-        setNewTodo(e.target.value);
-    };
-    return (
-        <div>
-            {!state.editing ? (
-                <h1>
-            {state.todo}{''}
-            <i
-            onClick={() => dispatch({type: 'ADD_TODO'})}
-            className='far fa-edit'
-            />
-            </h1>
-            ) : (
-                <div>
-            <input>
-            className='todo-input'
-            type='text'
-            name='newTodoText'
-            value={newTodoText}
-            onChange={handleChanges}
-            
-            </input>
-        </div>
-        <button
-        onClick={() =>
-            dispatch({ type: 'MARK_TODO', payload: newTodoText})
-        }
-        >
-         Mark Todo
-        </button>
-        </div>
-    );
 
-};
 
-export default Reducer;
+export default reducer;
